se incluyen cambios para mostrar el mapa
